Version 4.0.6
https://www.gurux.fi/node/15711

+++ development/src/main/java/gurux/dlms/objects/GXDLMSActivityCalendar.java
@@ -82,7 +82,7 @@ public class GXDLMSActivityCalendar extends GXDLMSObject
      *            Logical Name of the object.
      */
     public GXDLMSActivityCalendar(final String ln) {
-        this(ln, (short) 0);
+        this(ln, 0);
     }
 
     /**
