+++ development/src/main/java/gurux/dlms/objects/GXDLMSAutoConnect.java
@@ -84,7 +84,7 @@ public class GXDLMSAutoConnect extends GXDLMSObject implements IGXDLMSBase {
      *            Logical Name of the object.
      */
     public GXDLMSAutoConnect(final String ln) {
-        this(ln, (short) 0);
+        this(ln, 0);
     }
 
     /**
